RViz marker's need a unique namespace + id
Markers in have a lifetime and can be manipulated. The default action is
"add/modify" meaning it will add a new marker with the given namespace
or id if one doesn't already exist, otherwise it will modify the
existing one.

This commit makes the 3 axes have different id's so that they can all be
shown at once.